feat: add asset automatic
