+++ frontend/pages/home/components/ICRC/asset/AssetMutate.tsx
@@ -1,20 +1,17 @@
 import PlusIcon from "@assets/svg/files/plus-icon.svg";
 // svgs
-import AddAssetDrawer from "./AddAssetDrawer";
+import AssetDrawer from "./AssetDrawer";
 import { AssetMutationAction, setAssetMutationAction } from "@redux/assets/AssetReducer";
 import { useAppDispatch } from "@redux/Store";
 
-export default function AddAsset() {
+export default function AssetMutate() {
   const dispatch = useAppDispatch();
   return (
     <>
-      <div
-        className="flex flex-row items-center justify-center w-8 h-8 rounded-md cursor-pointer bg-SelectRowColor"
-        onClick={onAddAsset}
-      >
+      <div className={containerStyles} onClick={onAddAsset}>
         <img src={PlusIcon} alt="plus-icon" />
       </div>
-      <AddAssetDrawer />
+      <AssetDrawer />
     </>
   );
 
@@ -22,3 +19,5 @@ export default function AddAsset() {
     dispatch(setAssetMutationAction(AssetMutationAction.ADD_AUTOMATIC));
   }
 }
+
+const containerStyles = "flex flex-row items-center justify-center w-8 h-8 rounded-md cursor-pointer bg-SelectRowColor";
